+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/UserConnectionsCommander.scala
@@ -117,29 +117,6 @@ class UserConnectionsCommander @Inject() (
     }
   }
 
-  def getFriendRecommendations(userId: Id[User], offset: Int, limit: Int): Future[Option[FriendRecommendations]] = {
-    abookServiceClient.getFriendRecommendations(userId, offset, limit).map {
-      _.map { recommendedUsers =>
-        val friends = db.readOnlyReplica { implicit session =>
-          (recommendedUsers.toSet + userId).map(id => id -> userConnectionRepo.getConnectedUsers(id)).toMap
-        }
-
-        val mutualFriends = recommendedUsers.map { recommendedUserId =>
-          recommendedUserId -> (friends(userId) intersect friends(recommendedUserId))
-        }.toMap
-
-        val (basicUsers, mutualFriendConnectionCounts) = db.readOnlyReplica { implicit session =>
-          val uniqueMutualFriends = mutualFriends.values.flatten.toSet
-          val basicUsers = basicUserRepo.loadAll(uniqueMutualFriends ++ recommendedUsers)
-          val mutualFriendConnectionCounts = uniqueMutualFriends.map { mutualFriendId => mutualFriendId -> userConnectionRepo.getConnectionCount(mutualFriendId) }.toMap
-          (basicUsers, mutualFriendConnectionCounts)
-        }
-
-        FriendRecommendations(basicUsers, mutualFriendConnectionCounts, recommendedUsers, mutualFriends)
-      }
-    }
-  }
-
   def getMutualFriends(user1Id: Id[User], user2Id: Id[User]): Set[Id[User]] = {
     val (user1FriendIds, user2FriendIds) = db.readOnlyReplica { implicit session =>
       (userConnectionRepo.getConnectedUsers(user1Id), userConnectionRepo.getConnectedUsers(user2Id))
