+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/website/PeopleRecommendationController.scala
@@ -1,6 +1,6 @@
 package com.keepit.controllers.website
 
-import com.keepit.commanders.{ InviteCommander, UserConnectionsCommander }
+import com.keepit.commanders.{ UserCommander, InviteCommander, UserConnectionsCommander }
 import com.keepit.common.controller.{ UserActions, UserActionsHelper, ShoeboxServiceController }
 import com.google.inject.Inject
 import com.keepit.abook.ABookServiceClient
@@ -21,7 +21,7 @@ class PeopleRecommendationController @Inject() (
     basicUserRepo: BasicUserRepo,
     userRepo: UserRepo,
     db: Database,
-    peopleRecoCommander: UserConnectionsCommander,
+    peopleRecoCommander: UserCommander,
     socialUserRepo: SocialUserInfoRepo,
     inviteCommander: InviteCommander) extends UserActions with ShoeboxServiceController {
