+++ kifi-backend/modules/shoebox/test/com/keepit/commanders/UserConnectionsCommanderTest.scala
@@ -2,7 +2,9 @@ package com.keepit.commanders
 
 import com.keepit.abook.{ FakeABookServiceClientModule, FakeABookServiceClientImpl, ABookServiceClient }
 import com.keepit.common.db.Id
+import com.keepit.common.social.FakeSocialGraphModule
 import com.keepit.model.{ UserRepo, User, UserConnectionRepo, UserConnection }
+import com.keepit.scraper.FakeScrapeSchedulerModule
 import com.keepit.test.ShoeboxTestInjector
 import org.specs2.mutable.Specification
 
@@ -12,7 +14,9 @@ import concurrent.duration.Duration
 class UserConnectionsCommanderTest extends Specification with ShoeboxTestInjector {
 
   val modules = Seq(
-    FakeABookServiceClientModule()
+    FakeABookServiceClientModule(),
+    FakeScrapeSchedulerModule(),
+    FakeSocialGraphModule()
   )
 
   "PeopleRecommendationCommander" should {
@@ -23,7 +27,7 @@ class UserConnectionsCommanderTest extends Specification with ShoeboxTestInjecto
           val abook = inject[ABookServiceClient].asInstanceOf[FakeABookServiceClientImpl]
           abook.addFriendRecommendationsExpectations(users(0).id.get, Seq(users(1).id.get, users(2).id.get, users(3).id.get))
 
-          val friendRecoDataF = inject[UserConnectionsCommander].getFriendRecommendations(users(0).id.get, 2, 25)
+          val friendRecoDataF = inject[UserCommander].getFriendRecommendations(users(0).id.get, 2, 25)
           val friendRecoData = Await.result(friendRecoDataF, Duration(5, "seconds")).get
 
           friendRecoData.basicUsers === Map(
